Re-structure the unit tests
