+++ unit_tests/indexing_tests/test_data.py
@@ -1,0 +1,48 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+# Copyright (C) 2022 University of Dundee & Open Microscopy Environment.
+# All rights reserved.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+containers_1 = {
+    "1273": {
+        "type": "project",
+        "name": "idr0021-R",
+        "image count": 414,
+        "data_source": "omero1",
+    },
+    "701": {
+        "type": "project",
+        "name": "ggg",
+        "image count": 1,
+        "data_source": "omero1",
+    },
+}
+
+containers_2 = {
+    "1582": {
+        "type": "project",
+        "name": "matlab-project",
+        "image count": 4,
+        "data_source": "omero1",
+    },
+    "1270": {
+        "type": "project",
+        "name": "idr0021",
+        "image count": 33,
+        "data_source": "omero1",
+    },
+}
