+++ unit_tests/queries_tests/test_quries.py
@@ -28,12 +28,6 @@ from omero_search_engine.api.v1.resources.utils import (
     elasticsearch_query_builder,
     search_resource_annotation,
     get_data_sources,
-    update_data_source_cache,
-    delete_container,
-)
-
-from omero_search_engine.api.v1.resources.resource_analyser import (
-    return_containes_images,
 )
 
 from omero_search_engine.cache_functions.elasticsearch.elasticsearch_templates import (  # noqa
@@ -50,7 +44,6 @@ from omero_search_engine.cache_functions.elasticsearch.transform_data import (
     delete_es_index,
     create_index,
     get_all_indexes_from_elasticsearch,
-    index_container_s_from_database,
 )
 
 
@@ -72,8 +65,6 @@ from test_data import (
     image_owner,
     image_group,
     image_owner_group,
-    containers_1,
-    containers_2,
 )
 
 from omero_search_engine import search_omero_app, create_app
@@ -416,88 +407,6 @@ class BasicTestCase(unittest.TestCase):
         :return:
         """
         pass
-
-    def test_delete_index_one_container(self):
-        import time
-
-        # test delete container
-        for id, container in containers_1.items():
-            delete_container(id, container["type"], container["data_source"], False)
-            update_data_source_cache(container["data_source"])
-            time.sleep(20)
-            containers_ad = return_containes_images(
-                container["data_source"],
-            )
-            for con1 in containers_ad["results"]["results"]:
-                self.assertNotEquals(int(con1["id"]), int(id))
-
-        # test index container
-        resources_index = {
-            "project": ["image", "project"],
-            "screen": ["image", "screen", "well", "plate"],
-        }
-        for id, container in containers_1.items():
-            for res in resources_index[container["type"]]:
-                index_container_s_from_database(
-                    container["type"], res, id, container["data_source"]
-                )
-
-            update_data_source_cache(container["data_source"])
-            search_omero_app.logger.info("Waiting to index the data... ")
-
-            time.sleep(20)
-            # return_containes_images
-            containers_ai = return_containes_images(container["data_source"])
-
-            found = False
-            cur_res = None
-            for con1 in containers_ai["results"]["results"]:
-                if int(con1["id"]) == int(id) and con1["type"] == container["type"]:
-                    found = True
-                    cur_res = con1
-                    break
-            self.assertTrue(found)
-            self.assertEqual(int(cur_res["image count"]), int(container["image count"]))
-            self.assertEqual(cur_res["name"], container["name"])
-
-    def test_delete_index_multi_containers(self):
-        resources_index = {
-            "project": ["image", "project"],
-            "screen": ["image", "screen", "well", "plate"],
-        }
-        import time
-
-        ids = []
-        data_source = ""
-        resource = ""
-        for id in containers_2:
-            data_source = containers_2[id]["data_source"]
-            resource = containers_2[id]["type"]
-            ids.append(id)
-        delete_container(",".join(ids), resource, data_source, True)
-        time.sleep(30)
-        # return_containes_images
-        containers_ad = return_containes_images(data_source)
-
-        for id, container in containers_2.items():
-            for con1 in containers_ad["results"]["results"]:
-                self.assertNotEquals(int(con1["id"]), int(id))
-
-        for res in resources_index[resource]:
-            index_container_s_from_database(resource, res, ",".join(ids), data_source)
-        update_data_source_cache(data_source)
-        time.sleep(30)
-        containers_ai = return_containes_images(data_source)
-
-        for id, container in containers_2.items():
-            for con1 in containers_ai["results"]["results"]:
-                if int(con1["id"]) == int(id) and con1["type"] == container["type"]:
-                    found = True
-                    cur_res = con1
-                    break
-            self.assertTrue(found)
-            self.assertEqual(int(cur_res["image count"]), int(container["image count"]))
-            self.assertEqual(cur_res["name"], container["name"])
 
 
 if __name__ == "__main__":
